Converted notify_xmpp
